+++ cluster/pulumi/validator-runbook/src/installNode.ts
@@ -35,6 +35,7 @@ import {
   validatorSecrets,
   ValidatorTopupConfig,
   InstalledHelmChart,
+  ansDomainPrefix,
 } from 'splice-pulumi-common';
 import { installParticipant } from 'splice-pulumi-common-validator';
 import { SplicePostgres } from 'splice-pulumi-common/src/postgres';
@@ -111,6 +112,9 @@ export async function installNode(auth0Client: Auth0Client): Promise<void> {
       cluster: {
         hostname: CLUSTER_HOSTNAME,
         svNamespace: RUNBOOK_NAMESPACE,
+      },
+      spliceDomainNames: {
+        nameServiceDomain: ansDomainPrefix,
       },
       withSvIngress: false,
     },